pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "portal-inmobiliario-scraper"
version = "0.1.0"
description = "Production-ready scraper for Portal Inmobiliario Chile"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.44.0",
    "selectolax>=0.3.17",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "typer>=0.12.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "pyyaml>=6.0",
    "structlog>=24.0.0",
    "tenacity>=8.2.0",
    "schedule>=1.2.0",
    "pandas>=2.1.0",
    "pyarrow>=14.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "python-slugify>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
]

[project.scripts]
piscraper = "src.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v"

[tool.ruff]
line-length = 100
target-version = "py311"
